Fix the embedded overlay prototype example for on limb function

+++ NPC Maker/NPC Maker/Resources/EmbeddedOverlay.c
@@ -22,7 +22,7 @@ float NpcM_DoAThing(NpcMaker* npc, PlayState* play)
 
 The exception for this is functions which should run "On limb", their signature is as follows:
 
-int NpcM_Function(NpcMaker* npc, PlayState* playState, s32 limbNumber, Gfx** dListPtr, Vec3f* translation, Vec3s* rotation, void* instance, Gfx** gfxP)
+float NpcM_Function(NpcMaker* npc, PlayState* playState, s32 limbNumber, Gfx** dListPtr, Vec3f* translation, Vec3s* rotation, void* instance, Gfx** gfxP)
 
 This function should return 0 if the limb is meant to be drawn, and 1 if it is not.
